Fix escaped newline characters in JS
When a string is wrapped in single quotes but contains a JS escape sequence (like `\n`) and run through `esc_js()`, the slashes get stripped instead of converted to their proper sequence. In this case, that means a message that contains `dashboard.nnAre you sure` instead of a line break.

Converting the string quotes to double-quotes enables the escape sequence as expected.